Add approved version of the term sheet
This commit adds the version of the term sheet that was formally approved
by the Thunderbird council on 2022-08-01.

The original document used bullet points only for the first level of
hierarchy, while letters and Roman numerals were used for subsequent
levels. As markdown does not support the latter two, this representation
of the document is using bullet points on all levels.